Dockerizing MediaGoblin, working test image.
Dockerfile installs everything that's necessary for a base test setup.
Testing is done running the container interactively. Attempted to run
the container with '-d', but the email verification link wasn't showing
in the logs. Logs seemed a mess with '-d'.

So far I've only tested uploading audio files and a single image.
I attemtped to use the 1.0 versions of
python-gst0.10 and gstreamer0.10-plugins-{base,bad,good,ugly},
but that resulted in errors.

http://mediagoblin.readthedocs.org/en/latest/siteadmin/media-types.html
says to use gstreamer0.10-ffmpeg, but on Ubuntu 14.04, gstreamer1.0-libav
is available instead. That works without problems, as far as audio goes.

I'm not sure how best to setup Postgres, I'm aiming to set it up as a
separate container that this one links to. For testing, it's setup in
the same container.