Remove 'audio' plugin.
`scikits.audiolab` does not seem to get installed correctly, causing
the `lazyserver.sh` script to fail immediately. Attempting to import
scikits.audiolab in an interpreter using the python in a container
started with `--entrypoint /bin/bash` resulted in `version` not being
found despite there being a version.py file in the audiolab directory.

Removing the config option to see if there are any other issues.