`--server-name=broadcast` is invalid. What?
`lazyserver.sh` does not seem setup to handle `--server-name` as an
argument. Either the documentation has not been updated to reflect this,
or I'm missing something (hopefully) obvious. Removing the parameter,
and adding a paste_local.ini config; this seems to be a new config that
I don't remember from when I initially set things up.

There looks to be an appropraite `broadcast` config in the
`paste_local.ini` file, but I don't know how to use it.

`lazyserver.sh` seems to complain about a lack of a `mediagoblin.ini`
file if one is not present. Updating the Dockerfile to ADD
`mediagoblin_local.ini` as `mediagoblin.ini`. /shrug